+++ environments/nemo_gym/nemo_math_advanced_calculations/nemo_math_advanced_calculations.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 import verifiers as vf
-from verifiers.envs.integrations.nemo_gym_env import (
+from verifiers.envs.integrations.nemo_gym import (
     NemoGymEnv,
     _build_dataset,
     _reward_from_verify,
@@ -10,15 +10,15 @@ from verifiers.envs.integrations.nemo_gym_env import (
 
 def load_environment(
     dataset_split: str = "example",
-    max_turns: int = 16,
+    max_turns: int = 8,
     **kwargs: Any,
 ) -> vf.Environment:
     dataset, _ = _build_dataset(
-        resource_server="workplace_assistant", dataset_split=dataset_split
+        resource_server="math_advanced_calculations", dataset_split=dataset_split
     )
     rubric = vf.Rubric(funcs=[_reward_from_verify], weights=[1.0])
     return NemoGymEnv(
-        resource_server="workplace_assistant",
+        resource_server="math_advanced_calculations",
         dataset=dataset,
         rubric=rubric,
         max_turns=max_turns,
