+++ environments/nemo_gym/nemo_math_with_code/nemo_math_with_code.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 import verifiers as vf
-from verifiers.envs.integrations.nemo_gym_env import (
+from verifiers.envs.integrations.nemo_gym import (
     NemoGymEnv,
     _build_dataset,
     _reward_from_verify,
@@ -14,13 +14,15 @@ def load_environment(
     **kwargs: Any,
 ) -> vf.Environment:
     dataset, _ = _build_dataset(
-        resource_server="example_multi_step", dataset_split=dataset_split
+        resource_server="math_with_code", dataset_split=dataset_split
     )
     rubric = vf.Rubric(funcs=[_reward_from_verify], weights=[1.0])
     return NemoGymEnv(
-        resource_server="example_multi_step",
+        resource_server="math_with_code",
         dataset=dataset,
         rubric=rubric,
         max_turns=max_turns,
+        config_overrides={"max_execution_time": 10},
+        extra_pip_packages=["numpy", "scipy", "pandas"],
         **kwargs,
     )
