+++ environments/nemo_gym/nemo_xlam_fc/nemo_xlam_fc.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 import verifiers as vf
-from verifiers.envs.integrations.nemo_gym_env import (
+from verifiers.envs.integrations.nemo_gym import (
     NemoGymEnv,
     _build_dataset,
     _reward_from_verify,
@@ -10,19 +10,14 @@ from verifiers.envs.integrations.nemo_gym_env import (
 
 def load_environment(
     dataset_split: str = "example",
-    max_turns: int = 8,
     **kwargs: Any,
 ) -> vf.Environment:
-    dataset, _ = _build_dataset(
-        resource_server="math_with_code", dataset_split=dataset_split
-    )
+    dataset, _ = _build_dataset(resource_server="xlam_fc", dataset_split=dataset_split)
     rubric = vf.Rubric(funcs=[_reward_from_verify], weights=[1.0])
     return NemoGymEnv(
-        resource_server="math_with_code",
+        resource_server="xlam_fc",
         dataset=dataset,
         rubric=rubric,
-        max_turns=max_turns,
-        config_overrides={"max_execution_time": 10},
-        extra_pip_packages=["numpy", "scipy", "pandas"],
+        max_turns=1,
         **kwargs,
     )
